requirement.txt
flask
flask-cors
openai-whisper
transformers
Pillow
librosa
exifread
torch